--- a/lib/pipeline.js
+++ b/lib/pipeline.js
@@ -864,8 +864,10 @@ class PipelineModel extends BaseModel {
                         job.archived = false;
                         updatedJobs.push(await job.update());
                     } else if (!job.isPR()) {
-                        job.archived = true;
-                        updatedJobs.push(await job.update());
+                        if (!job.archived) {
+                            job.archived = true;
+                            updatedJobs.push(await job.update());
+                        }
                     }
 
                     // sync external triggers for existing jobs