## Context

<!-- Why do we need this PR? What was the reason that led you to make this change? -->
`v4/events` performance issue makes problem like https://github.com/screwdriver-cd/screwdriver/issues/2318
We tried to fix it before with the following PR, but it still seems a little slow.
https://github.com/screwdriver-cd/models/pull/485

When we looked into it, we found a process that was doing an unnecessary UPDATE.
https://github.com/screwdriver-cd/models/blob/3a2dbd03a42dda7b7b0bbaaf09d2d72c116fdb01/lib/pipeline.js#L867-L868

The following is an example of a change in `v4/events` velocity.
However, the speed will be affected by the network speed and other factors.
In my sandbox, it looks like this
```
Archived jobs : 1000
Before: 9.61s
After: 7.00s
```
```
Archived jobs : 4000
Before: 16.63s
After: 14.53s
```

It looks like it will perform a little better.
## Objective

<!-- What does this PR fix? What intentional changes will this PR make? -->
If archived is already true, do not update it.

## References

<!-- Links or resources that help clarify and support your intentions (e.g., Github issue) -->
- https://github.com/screwdriver-cd/models/pull/485
- https://github.com/screwdriver-cd/screwdriver/issues/2318
- https://github.com/screwdriver-cd/screwdriver/issues/2311
## License

<!-- The following line must be included in your pull request -->

I confirm that this contribution is made under the terms of the license found in the root directory of this repository's source tree and that I have the authority necessary to make this contribution on behalf of its copyright owner.