@yoshwata 2021-02-15
When I checked before it spends over 80000ms, so 1000ms or a few seconds seems to be very small. It might not be considered an effective measure.

@jithine We have some issues with a large number of archived jobs. Cpu utilization increases significantly when `/events` in a pipeline with a large number of archived jobs are executed. Does Verizon take measures such as deleting archived jobs on a regular basis?